feat: add function build in build.sh

+++ build.sh
@@ -1,5 +1,13 @@
 #!/bin/sh
-mkdir -p build
-rm -rf build/*
-cp target/release/lancelot_life build/server
-cp -r static build/
+function build() {
+  if cargo build --release
+  then
+    mkdir -p build
+    rm -rf build/*
+    cp target/release/lancelot_life build/server
+    cp -r static build/
+  else
+    echo "build.sh: Build Failed with code $?"
+    exit 1
+  fi
+}
